CMC show DoB question for both actors
