+++ app/journeys/cmc.js
@@ -1,10 +1,6 @@
 'use strict';
 
-const getJourney = (actor) => {
-    return {stepList: actor === 'claimant' ? claimant : defendant};
-};
-
-const claimant = {
+const stepList = {
     StartPage: 'ApplicantDateOfBirth',
     ApplicantDateOfBirth: 'ApplicantLanguage',
     ApplicantLanguage: {
@@ -42,41 +38,6 @@ const claimant = {
     ApplicantPregnant: 'EndPage'
 };
 
-const defendant = {
-    StartPage: 'ApplicantLanguage',
-    ApplicantLanguage: {
-        otherLanguage: 'ApplicantEnglishLevel',
-        otherwise: 'ApplicantSex'
-    },
-    ApplicantEnglishLevel: 'ApplicantSex',
-    ApplicantSex: 'ApplicantGenderSameAsSex',
-    ApplicantGenderSameAsSex: 'ApplicantSexualOrientation',
-    ApplicantSexualOrientation: 'ApplicantMaritalStatus',
-    ApplicantMaritalStatus: 'ApplicantEthnicGroup',
-    ApplicantEthnicGroup: {
-        White: 'ApplicantEthnicBackgroundWhite',
-        Mixed: 'ApplicantEthnicBackgroundMixed',
-        Asian: 'ApplicantEthnicBackgroundAsian',
-        Black: 'ApplicantEthnicBackgroundBlack',
-        Other: 'ApplicantEthnicBackgroundOther',
-        otherwise: 'ApplicantReligion'
-    },
-    ApplicantEthnicBackgroundWhite: 'ApplicantReligion',
-    ApplicantEthnicBackgroundMixed: 'ApplicantReligion',
-    ApplicantEthnicBackgroundAsian: 'ApplicantReligion',
-    ApplicantEthnicBackgroundBlack: 'ApplicantReligion',
-    ApplicantEthnicBackgroundOther: 'ApplicantReligion',
-    ApplicantReligion: 'ApplicantDisability',
-    ApplicantDisability: {
-        Yes: 'ApplicantDisabilityImplications',
-        otherwise: 'ApplicantPregnant'
-    },
-    ApplicantDisabilityImplications: {
-        Yes: 'ApplicantDisabilityImplicationAreas',
-        otherwise: 'ApplicantPregnant'
-    },
-    ApplicantDisabilityImplicationAreas: 'ApplicantPregnant',
-    ApplicantPregnant: 'EndPage'
+module.exports = () => {
+    return {stepList};
 };
-
-module.exports = getJourney;
